+++ cleanlinks.sh
@@ -1,3 +1,5 @@
+sudo rm ~/penguins-eggs-/tmp -rf
+sudo rm ~/penguins-eggs-/dist -rf
 
 sudo rm /usr/bin/add-penguins-links.sh
 sudo rm /usr/bin/penguins-links-add.sh
@@ -5,7 +7,8 @@ sudo rm /usr/bin/penguins-links-add.sh
 sudo rm /usr/sbin/before-bootloader.*
 sudo rm /usr/sbin/before-bootloader-mkdirs*
 sudo rm /usr/sbin/bug*
-suro rm /usr/sbin/after-bootloader*
+sudo rm /usr/sbin/after-bootloader*
+sudo rm /usr/sbin/remove-link.*
 
 sudo rm /usr/share/applications/penguins-*
 sudo rm /usr/share/applications/dw*
@@ -17,8 +20,6 @@ sudo rm /etc/xdg/autostart/add-penguins-links.desktop
 
 # moduli calamares
 sudo rm /etc/calamares -rf
-
-/usr/sbin/remove-link.sh
 
 sudo rm /usr/lib/x86_64-linux-gnu/calamares/modules/automirror/ -rf
 sudo rm /usr/lib/x86_64-linux-gnu/calamares/modules/bootloader-config/ -rf
@@ -35,3 +36,4 @@ sudo rm /usr/lib/calamares/modules/remove-link/ -rf
 sudo rm /usr/lib/calamares/modules/sources-final/ -rf
 sudo rm /usr/lib/calamares/modules/sources-trusted/ -rf
 sudo rm /usr/lib/calamares/modules/sources-trusted-unmount/ -rf
+