+++ src/classes/incubation/fisherman-helper/packages.ts
@@ -1,3 +1,4 @@
+import { O_APPEND } from 'constants'
 import { IDistro } from '../../../interfaces'
 /**
  *
@@ -7,13 +8,54 @@ import Pacman from '../../pacman'
 /**
  *
  */
-export function packages(distro: IDistro): string {
-   let text = ``
+export function remove(distro: IDistro): string {
+   let text = '  - remove:\n'
    text += removeEggs(distro)
    text += '\n'
    return text
 }
 
+/**
+ * 
+ * @param distro 
+   - try_install:
+      - language-pack-$LOCALE
+      - hunspell-$LOCALE
+      - libreoffice-help-$LOCALE
+
+ */
+export function tryInstall(distro: IDistro): string {
+   let text = `   - try_install:\n`
+
+
+   // Pacchetti da installare sempre
+   text += '    - hunspell-$LOCALE\n'
+
+   // Pacchetti da installare a seconda della distribuzione
+   if ((distro.versionLike === 'focal') || (distro.versionLike === 'bionic')) {
+      text += '    - language-pack-$LOCALE\n'
+   }
+
+   // Pacchetti da installare se sono presenti
+   if (Pacman.packageIsInstalled(`libreoffice-base-core`)) {
+      text += '    - libreoffice-l10n-$LOCALE\n'
+      text += '    - libreoffice-help-$LOCALE\n'
+   }
+
+   if (Pacman.packageIsInstalled('firefox-esr')) {
+      text += '    - firefox-esr-$LOCALE\n'
+   }
+
+   return text
+}
+
+
+
+
+/**
+ * 
+ * @param distro 
+ */
 function removeEggs(distro: IDistro): string {
    const packages = Pacman.packages()
    let text = ''